feat: added merging for free blocks into BumpAllocator
A merging system was added to the BumpAllocator to merge free blocks into a single block.

It serves for cases when we have a lot of small free blocks and we want to allocate a single big block.
So we can merge all the free blocks into a single one and allocate from it.